+++ gmf_core/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Enhanced memory alignment support for GMF data_bus and payload
 - Add a breaking change for the `esp_gmf_fifo_set_align` api usage
 - Added `HOWL` capability definition
+- Added `DOA` capability definition (`ESP_GMF_CAPS_AUDIO_DOA`)
 
 ## v0.8.4